Minor tweaks
Changed the game's name in project settings, made a few tweaks to the new text output box, added a "walk" and a "grab" command in game.cs, they work exactly like "move" and "take", no conflicts. Added a sprite to a room on the map but it looks bad now.