Added help command and tooltips
Added a ‘help’ command to game.cs.
Added a tooltip for the map title.
Added a label node and a tooltip for the map marker.
Adjusted text box for Inventory title.
Added a tooltip for the Inventory title.
Added a selection colour theme override for the text output box.
Added a placeholder text prompt for the text input box.
Fixed various typos in strings relating to the font’s  ‘ symbol.
Text output now has a context menu (From earlier commit).